First Day
My brain is dead, it is 1am, and I have alot of framework code, just nothing polished. Towers can be placed in towerSlots, money system works (no way to gain). Need to code in enemies, and their spawning, and then define values for everything, also health system needs to be added, and a winZone for the enemies, and a path system. ARGH